Split age requirement function into two functions
Introduce `meetsMinimumAgeRequirement()` and `meetsMaximumAgeRequirement()` to replace `meetsAgeRequirements()`.
The two functions allow us to show a different error message if a given date is too recent or too far back. For
the case a given date is too old for an age, we're reusing the existing invalid date error message for simplicity.