AttachmentCarousel: Throw error when an image can't be mapped
Enhances error handling in AttachmentCarousel to provide a clearer user experience when an image cannot be mapped to available content. Previously, if the carousel failed to find the correct image to display, it defaulted to showing the first image. This behaviour lead to user confusion as the displayed image did not correspond to the selected one. With this change, if a user tries to open an image that cannot be mapped, an error is thrown and an error page is displayed. This explicit error handling helps users understand when there's a problem opening an image, prompting them to contact support or attempt to view another image.